+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 
 ## 1.1.1 under development
 
-- no changes in this release.
+- Enh #70: Add support `\Stringable` as content in methods `Html::tag()`, `Html::normalTag()`, `Html::a()`,
+  `Html::label()`, `Html::option()`, `Html::div()`, `Html::span()`, `Html::p()`, `Html::li()`, `Html::caption()`,
+  `Html::td()`, `Html::th()` (vjik)
 
 ## 1.1.0 April 09, 2021